+++ src/components/Breadcrumb/breadcrumb.story.jsx
@@ -1,26 +1,26 @@
 import React from 'react';
-import Breadcrumb from './index';
+import Breadcrumb from './Breadcrumb';
 
 const mockedItems = [
-	{
-		link: ' ',
-		text: 'breadcrumb 1',
-	},
-	{
-		link: ' ',
-		text: 'breadcrumb 2',
-	},
-	{
-		text: 'breadcrumb 3',
-	},
+    {
+        link: ' ',
+        text: 'breadcrumb 1',
+    },
+    {
+        link: ' ',
+        text: 'breadcrumb 2',
+    },
+    {
+        text: 'breadcrumb 3',
+    },
 ];
 
 export default {
-	title: 'Breadcrumb',
+    title: 'Breadcrumb',
 };
 
 export const Default = () => (
-	<div>
-		<Breadcrumb items={mockedItems} />
-	</div>
+    <div>
+        <Breadcrumb items={mockedItems} />
+    </div>
 );
